Fixing stddev computation ?
It seems that this -1 may not be suitable. Im not sureThis is acording to my low level in stats. but a check on the internet seems to agree. [link on kahnacademy](https://www.khanacademy.org/math/probability/data-distributions-a1/summarizing-spread-distributions/a/calculating-standard-deviation-step-by-step)

 Im not even sure. i havent even tested the code. 

ps: thanks for the work on tensorflow.js/deeplearnjs, i having a lot of fun learning it :)