Upgrade six node-based examples  (#366)
Upgrade the following examples to tfjs 1.7.4:
- date-conversion-attention
- fashion-mnist
- mnist-acgan
- sentiment
- snake-dqn

Full upgrade to 2.6.0 is blocked in these examples due to ES6 module changes in tfjs, which breaks the unit tests. This issue will be addressed later.

Upgrade the following example to tfjs 2.6.0:
- simple-object-detection

(This example doesn't have unit tests yet.)

In addition, in the date-conversion-attention example, fix a bug in the sorting of the x- and y-label ticks that causes the rows and columns of the attention matrix to go out of order.

All six examples are tested manually to ensure that they continue to work.