Update and simplify snake_game.js game engine (2 liner)
Hi there!

I noted that the snake died a little early in some longer games, so I reworked the game engine to allow for longer snakes. It's now simplified to an additional array that is run through to see if the snake head has already been there. Thank you for your work!